pyproject.toml
[tool.poetry]
name = "render"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^21.2.0"
Pillow = "^10.1.0"
PyJWT = "^2.8.0"
pytz = "^2023.3.post1"
whitenoise = "^6.6.0"
django-environ = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
